RL/requirements.txt
#--extra-index-url https://download.pytorch.org/whl/cpu
pytorch
numpy
#opencv-python
gymnasium
#stable_baselines3
